Move general linter rendering to VersionFileViewer
