+++ src/components/App/styles.module.scss
@@ -28,12 +28,3 @@
     margin-bottom: 0;
   }
 }
-
-.linterMessagesShell {
-  padding: $default-padding;
-  padding-bottom: 0;
-}
-
-.linterMessage:last-child {
-  margin-bottom: 0;
-}
