+++ src/components/FullscreenGrid/index.tsx
@@ -21,6 +21,7 @@ type Props = PublicProps & PropsFromState;
 export enum PanelAttribs {
   altSidePanel = 'altSidePanel',
   mainSidePanel = 'mainSidePanel',
+  topContent = 'topContent',
 }
 
 export const Header = ({
