+++ src/components/VersionFileViewer/index.spec.tsx
@@ -9,9 +9,13 @@ import FileMetadata from '../FileMetadata';
 import FileTree from '../FileTree';
 import { PanelAttribs } from '../FullscreenGrid';
 import KeyboardShortcuts from '../KeyboardShortcuts';
+import LinterMessage from '../LinterMessage';
 import LinterProvider from '../LinterProvider';
 import Loading from '../Loading';
-import { getMessageMap } from '../../reducers/linter';
+import {
+  LinterMessage as LinterMessageType,
+  getMessageMap,
+} from '../../reducers/linter';
 import {
   ExternalVersionEntry,
   actions as versionsActions,
@@ -88,7 +92,7 @@ describe(__filename, () => {
 
   type RenderParams = Partial<PublicProps>;
 
-  const render = (moreProps: RenderParams) => {
+  const render = (moreProps: RenderParams = {}) => {
     const { file, version } = getInternalVersionAndFile();
     const props = {
       children: <div />,
@@ -101,9 +105,22 @@ describe(__filename, () => {
     return shallow(<VersionFileViewer {...props} />);
   };
 
-  const renderPanel = (params: RenderParams, panel: PanelAttribs) => {
+  const renderPanel = (params: RenderParams = {}, panel: PanelAttribs) => {
     const root = render(params);
     return getContentShellPanel(root, panel);
+  };
+
+  const renderTopContent = (params: RenderParams = {}) => {
+    return renderPanel(params, PanelAttribs.topContent);
+  };
+
+  const renderWithLinterMessages = (
+    root: ShallowWrapper,
+    messages: Partial<LinterMessageType>[] = [],
+  ) => {
+    return simulateLinterProvider(root, {
+      messageMap: getMessageMap(createFakeExternalLinterResult({ messages })),
+    });
   };
 
   const getItem = (root: ShallowWrapper, title: ItemTitles) => {
@@ -243,5 +260,72 @@ describe(__filename, () => {
 
     const overview = root.find(CodeOverview);
     expect(overview).toHaveProp('content', '');
+  });
+
+  it('configures LinterProvider', () => {
+    const { version } = getInternalVersionAndFile();
+    const root = renderTopContent({ version });
+
+    const provider = root.find(LinterProvider);
+    expect(provider).toHaveProp('versionId', version.id);
+    expect(provider).toHaveProp('validationURL', version.validationURL);
+    expect(provider).toHaveProp('selectedPath', version.selectedPath);
+  });
+
+  it('does not configure LinterProvider without a version', () => {
+    const root = render();
+
+    expect(root.find(LinterProvider)).toHaveLength(0);
+  });
+
+  it('renders general LinterMessages', () => {
+    const uid1 = 'general-message-1';
+    const uid2 = 'general-message-2';
+
+    const root = renderWithLinterMessages(renderTopContent(), [
+      {
+        file: null,
+        line: null,
+        uid: uid1,
+      },
+      {
+        file: null,
+        line: null,
+        uid: uid2,
+      },
+    ]);
+
+    const messages = root.find(LinterMessage);
+
+    expect(messages).toHaveLength(2);
+    expect(messages.at(0)).toHaveProp(
+      'message',
+      expect.objectContaining({ uid: uid1 }),
+    );
+    expect(messages.at(1)).toHaveProp(
+      'message',
+      expect.objectContaining({ uid: uid2 }),
+    );
+  });
+
+  it('does not render LinterMessages with an empty messageMap', () => {
+    const root = simulateLinterProvider(renderTopContent(), {
+      messageMap: undefined,
+    });
+
+    expect(root.find(LinterMessage)).toHaveLength(0);
+  });
+
+  it('ignores non-general LinterMessages', () => {
+    const root = renderWithLinterMessages(renderTopContent(), [
+      {
+        // Define a message for a file, which should be ignored.
+        file: 'some-file.js',
+        line: null,
+        uid: 'global-message-example',
+      },
+    ]);
+
+    expect(root.find(LinterMessage)).toHaveLength(0);
   });
 });
