+++ src/components/VersionFileViewer/styles.module.scss
@@ -1,5 +1,11 @@
+@import '../../scss/variables';
+
 .mainSidePanel,
 .altSidePanel {
   border: none;
   padding: 0;
+}
+
+.linterMessage:last-child {
+  margin-bottom: 0;
 }
